+++ src/parsley/factory/options.js
@@ -38,6 +38,9 @@ define('parsley/factory/options', [
     getFieldOptions: function (fieldInstance) {
       this.fieldOptions = ParsleyUtils.attr(fieldInstance.$element, this.staticOptions.namespace);
 
+      if (null === this.formOptions && 'ParsleyForm' === fieldInstance.parsleyInstance.__proxy__)
+        this.formOptions = getFormOptions(fieldInstance.parsleyInstance);
+
       // not deep extend, since formOptions and fieldOptions is a 1 level deep object
       return $.extend({}, this.staticOptions, this.formOptions, this.fieldOptions);
     }