+++ src/parsley/ui.js
@@ -247,10 +247,17 @@ define('parsley/ui', [
     },
 
     _insertErrorWrapper: function (fieldInstance) {
-      if ('string' === typeof fieldInstance.options.errorsContainer && $(fieldInstance.options.errorsContainer + '').length)
-        return $(fieldInstance.options.errorsContainer).append(fieldInstance._ui.$errorsWrapper);
+      var $errorsContainer;
 
-      var $errorsContainer = fieldInstance.options.errorsContainer(fieldInstance);
+      if ('string' === typeof fieldInstance.options.errorsContainer )
+        if ($(fieldInstance.options.errorsContainer + '').length)
+          return $(fieldInstance.options.errorsContainer).append(fieldInstance._ui.$errorsWrapper);
+        else if (window.console && window.console.warn)
+          window.console.warn('The errors container `' + fieldInstance.options.errorsContainer + '` does not exist in DOM');
+
+
+      if ('function' === typeof fieldInstance.options.errorsContainer)
+        $errorsContainer = fieldInstance.options.errorsContainer(fieldInstance);
 
       if ('undefined' !== typeof $errorsContainer && $errorsContainer.length)
         return $errorsContainer.append(fieldInstance._ui.$errorsWrapper);